fix(middlewares/utils): :recycle:  move errorhandler class to utils folder
