+++ server/src/middleware/errorHandler.js
@@ -1,20 +1,5 @@
-class ErrorHandler extends Error {
-  constructor(statusCode, messages, err) {
-    super();
-    this.statusCode = statusCode;
-    this.message = messages;
-    this.error = err;
-  }
-
-  logError(error) {
-    console.log(error);
-  }
-}
-
-function handlerErrors(err, req, res, next) {
-  const { statusCode, message, error } = err;
-
-  err.logError(error);
+function handlerErrors(err, req, res) {
+  const { statusCode, message } = err;
 
   return res.status(statusCode).json({
     statusCode,
@@ -22,4 +7,4 @@ function handlerErrors(err, req, res, next) {
   });
 }
 
-export { ErrorHandler, handlerErrors };
+export default handlerErrors;
