+++ server/src/utils/handleError.js
@@ -1,0 +1,10 @@
+class ErrorHandler extends Error {
+  constructor(statusCode, messages, err) {
+    super();
+    this.statusCode = statusCode;
+    this.message = messages;
+    this.error = err;
+  }
+}
+
+export default ErrorHandler;
